DjangoRESTapi/requirements.txt

# Django Cors Headers - Used to enable CORS headers in API responses, and allow requests to be made to your API server from other origins.
django-cors-headers==3.5.0

# Django Rest Framework - Used to design API logic
djangorestframework==3.12.2

# MySQL Client - Used as an interface to connect Django application to the MySQL server.
mysqlclient==2.1.1